Move Sample UI
