+++ gradle/libs.versions.toml
@@ -5,20 +5,19 @@ kotlin = "1.9.10"
 compose = "1.5.4"
 compose-compiler = "1.5.3"
 compose-material3 = "1.1.2"
-androidx-activityCompose = "1.8.0"
+androidx-activityCompose = "1.8.1"
 kotlinpoet = "1.14.2"
 kotlinxCoroutines = "1.7.3"
 ksp = "1.9.10-1.0.13"
-ktlint = "11.3.2"
+ktlint = "11.6.1"
 mockk = "1.13.8"
 multiplatformSettings = "1.1.0"
 preferenceKtx = "1.2.1"
 robolectric = "4.11.1"
 startupRuntime = "1.1.1"
-trikot = "5.3.0-dev2689"
+trikot = "5.3.0-dev2693"
 monitor = "1.6.1"
 miregoPublish = "1.0"
-miregoRelease = "2.0"
 
 [libraries]
 androidx-junit = { module = "androidx.test.ext:junit", version.ref = "junit" }
