+++ sample/common/src/commonMain/kotlin/com/mirego/debugpanel/sample/usecase/SampleDebugPanelUseCasePreview.kt
@@ -2,7 +2,6 @@ package com.mirego.debugpanel.sample.usecase
 
 import com.mirego.debugpanel.config.DebugPanelPickerItem
 import com.mirego.debugpanel.sample.Language
-import com.mirego.debugpanel.usecase.DebugPanelItemViewData
 import com.mirego.debugpanel.usecase.DebugPanelUseCasePreview
 import com.mirego.debugpanel.usecase.DebugPanelViewData
 import com.mirego.debugpanel.usecase.SampleDebugPanelUseCase
@@ -18,36 +17,7 @@ class SampleDebugPanelUseCasePreview : SampleDebugPanelUseCase, DebugPanelUseCas
         firstName: Flow<String>,
         environments: List<DebugPanelPickerItem>,
         resetOnboarding: () -> Unit
-    ): DebugPanelViewData = DebugPanelViewData(
-        listOf(
-            DebugPanelItemViewData.Toggle(
-                identifier = "toggle",
-                label = "Toggle",
-                initialValue = true
-            ),
-            DebugPanelItemViewData.TextField(
-                identifier = "textField",
-                placeholder = "Text field",
-                initialValue = ""
-            ),
-            DebugPanelItemViewData.Label(
-                identifier = "label",
-                label = "Label:",
-                value = flowOf("Value")
-            ),
-            DebugPanelItemViewData.Picker(
-                identifier = "picker",
-                label = "Picker",
-                initialValue = "",
-                items = listOf(DebugPanelPickerItem("id", "Item"))
-            ),
-            DebugPanelItemViewData.Button(
-                identifier = "button",
-                label = "Tap me",
-                action = {}
-            )
-        )
-    )
+    ): DebugPanelViewData = createViewData()
 
     override fun getPreviewMode(): Flow<Boolean?> = flowOf(null)
